added custom stab file

+++ oe/recipes/base-files/base-files_3.0.14.bb
@@ -118,7 +118,12 @@ do_install () {
                         install -m 0644 ${WORKDIR}/issue.net ${D}${sysconfdir}/issue.net
                 fi
 
-                install -m 0644 ${WORKDIR}/fstab ${D}${sysconfdir}/fstab
+		#MobiAqua: added handle custom fstab file
+		if [ -f ${MA_FSTAB_FILE} ]; then
+			install -m 0644 ${MA_FSTAB_FILE} ${D}${sysconfdir}/fstab
+		else
+			install -m 0644 ${WORKDIR}/fstab ${D}${sysconfdir}/fstab
+		fi
         	install -m 0644 ${WORKDIR}/filesystems ${D}${sysconfdir}/filesystems
         	install -m 0644 ${WORKDIR}/usbd ${D}${sysconfdir}/default/usbd
         	install -m 0644 ${WORKDIR}/profile ${D}${sysconfdir}/profile
